requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
baidupcsapi==0.9.4
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.1
hyperlink==19.0.0
idna==2.8
img2pdf==0.3.3
incremental==17.5.0
lxml==4.3.4
mysqlclient==1.4.2.post1
parsel==1.5.1
pdfkit==0.6.1
Pillow==6.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyPdf==1.13
PyPDF2==1.26.0
pytz==2019.1
pywin32==224
queuelib==1.5.0
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
Scrapy==1.7.2
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.2
Twisted==19.2.1
urllib3==1.25.3
virtualenv==16.7.1
w3lib==1.20.0
zope.interface==4.6.0
